Improvements for the python files
Improved the various python files by adding the ability to use an input file of the same format as vars.ini. Also added a script called alpha_tester.py which is used to plot L and check the validity of alpha values, i.e. if it maximizes for the right velocity distribution.